Initialized project
This initialized project contains the frontend and backend of The Pub project. Both will have their volumes mounted to the instantiated container for easy development.
Currently, this project is setup only for development. There are considerations when it comes to deploying live:
ASGI and WSGI for the Django backend
Building the frontend and running the start command instead of dev command

Keep this in mind for future production deployments.